Refactor unit tests for all events with all definitions (same test file)
